Return tokenized minutes and seconds as integers
There is no way it's useful to get a fraction of a second here since this is such a rough estimate. I don't even know if there's a point in having float accessors at all. Maybe internal protected one for the minute one.

See https://github.com/hydephp/develop/pull/963/files#r1103014593

+++ packages/framework/tests/Feature/ReadingTimeTest.php
@@ -78,11 +78,11 @@ class ReadingTimeTest extends TestCase
     public function test_formatUsingClosure()
     {
         /**
-         * @param  float  $minutes
-         * @param  float  $seconds
+         * @param  int  $minutes
+         * @param  int  $seconds
          * @return string
          */
-        $closure = function (float $minutes, float $seconds): string {
+        $closure = function (int $minutes, int $seconds): string {
             return "$minutes minutes, $seconds seconds";
         };
